codage du test processExitingVehicleTest qui teste si le véhicule est bien sorti et que la place qu'il occupait est à présent libre et qu'il a bien payé le bon prix si par exemple c'est un véhicule régulier

+++ parkingsystem/src/main/java/com/parkit/parkingsystem/dao/TicketDAO.java
@@ -55,7 +55,7 @@ public class TicketDAO {
                 ticket.setParkingSpot(parkingSpot);
                 ticket.setId(rs.getInt(2));
                 ticket.setVehicleRegNumber(vehicleRegNumber);
-                ticket.setPrice(rs.getDouble(3));
+                ticket.setPrice(rs.getFloat(3));
                 ticket.setInTime(rs.getTimestamp(4));
                 ticket.setOutTime(rs.getTimestamp(5));
             }
