+++ parkingsystem/src/main/java/com/parkit/parkingsystem/service/FareCalculatorService.java
@@ -14,7 +14,7 @@ public class FareCalculatorService {
         long outHour = ticket.getOutTime().getTime();
 
         //TODO: Some tests are failing here. Need to check if this logic is correct
-        float duration = (outHour - inHour)/3600000f;
+        double duration = (outHour - inHour)/3600000.0;
 
         switch (ticket.getParkingSpot().getParkingType()){
             case CAR: {
